Extract mapBriefToKoreanModernCStenoKeys
